+++ src/index.ts
@@ -1,16 +1,21 @@
 export {
-  ORG_SPONSOR_BUTTON_URL,
+  GithubSponsorButton,
+  type GithubSponsorButtonProps,
+  GithubSponsorButtonUrl,
   BtmtSponsorButton,
+  ORG_GITHUB_SPONSOR_BUTTON_URL,
   type BtmtSponsorButtonProps,
   Button,
   type ButtonProps,
   TheMusicTreeByline,
   TheMusicTreeHorizontalLink,
+  TheMusicTreeMarkLink,
   ORG_URL,
   parseOrgSiteHref,
   resolveOrgSiteHref,
   type TheMusicTreeBylineProps,
   type TheMusicTreeHorizontalLinkProps,
+  type TheMusicTreeMarkLinkProps,
 } from "./components";
 export * from "./tokens";
 export * from "./hooks";